+++ src/main/resources/application.yml
@@ -2,7 +2,7 @@ spring:
   profiles:
     active: ${SPRING_PROFILES_ACTIVE:dev} # 기본값 dev
   mail:
-    host: smtp.naver.com
+    host: smtp.gmail.com
     port: 465
     username: ${MAIL_USERNAME}
     password: ${MAIL_PASSWORD}
@@ -12,7 +12,7 @@ spring:
           auth: true
           ssl:
             enable: true
-            trust: smtp.naver.com
+            trust: smtp.gmail.com
         debug: true
 
 springdoc:
